LUI and ADDI working

+++ src/architecture.c
@@ -5,6 +5,6 @@
 #include "defines.h"
 
 mem MEMORY[MEM_SIZE] = {0};
-uint8_t databus = 0;
+uint32_t databus = 0;
 uint8_t PORTA = 0;
 uint8_t PORTB = 0;
